# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for ML Issues Dashboard
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Issue {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  priority: string;
  status: string;
  /** @nullable */
  responsible?: string | null;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateIssueInput {
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  priority: string;
  status: string;
  /** @nullable */
  responsible?: string | null;
  /** @nullable */
  dueDate?: string | null;
}

export interface UpdateIssueInput {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  responsible?: string | null;
  /** @nullable */
  dueDate?: string | null;
}

export interface IssueStats {
  total: number;
  open: number;
  inProgress: number;
  resolved: number;
  critical: number;
}

export interface CategoryCount {
  category: string;
  count: number;
  open: number;
  resolved: number;
}

export interface PriorityCount {
  priority: string;
  count: number;
}

export interface TrendPoint {
  date: string;
  created: number;
  resolved: number;
}

export type ListIssuesParams = {
  /**
   * @nullable
   */
  status?: string | null;
  /**
   * @nullable
   */
  category?: string | null;
  /**
   * @nullable
   */
  priority?: string | null;
};