# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for ML Issues Dashboard
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: issues
    description: Mercado Livre issue tracking
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /issues:
    get:
      operationId: listIssues
      tags: [issues]
      summary: List all issues
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: priority
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
    post:
      operationId: createIssue
      tags: [issues]
      summary: Create a new issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIssueInput"
      responses:
        "201":
          description: Created issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"

  /issues/stats:
    get:
      operationId: getIssueStats
      tags: [issues]
      summary: Get issue statistics
      responses:
        "200":
          description: Issue statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueStats"

  /issues/by-category:
    get:
      operationId: getIssuesByCategory
      tags: [issues]
      summary: Get issue counts grouped by category
      responses:
        "200":
          description: Issues by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryCount"

  /issues/by-priority:
    get:
      operationId: getIssuesByPriority
      tags: [issues]
      summary: Get issue counts grouped by priority
      responses:
        "200":
          description: Issues by priority
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriorityCount"

  /issues/trend:
    get:
      operationId: getIssuesTrend
      tags: [issues]
      summary: Get issues trend over time (last 30 days)
      responses:
        "200":
          description: Issues trend
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendPoint"

  /issues/{id}:
    get:
      operationId: getIssue
      tags: [issues]
      summary: Get a single issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Issue detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "404":
          description: Not found
    patch:
      operationId: updateIssue
      tags: [issues]
      summary: Update an issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIssueInput"
      responses:
        "200":
          description: Updated issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
    delete:
      operationId: deleteIssue
      tags: [issues]
      summary: Delete an issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Issue:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        priority:
          type: string
        status:
          type: string
        responsible:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - category
        - priority
        - status
        - createdAt
        - updatedAt

    CreateIssueInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        priority:
          type: string
        status:
          type: string
        responsible:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
      required:
        - title
        - category
        - priority
        - status

    UpdateIssueInput:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        responsible:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]

    IssueStats:
      type: object
      properties:
        total:
          type: integer
        open:
          type: integer
        inProgress:
          type: integer
        resolved:
          type: integer
        critical:
          type: integer
      required:
        - total
        - open
        - inProgress
        - resolved
        - critical

    CategoryCount:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
        open:
          type: integer
        resolved:
          type: integer
      required:
        - category
        - count
        - open
        - resolved

    PriorityCount:
      type: object
      properties:
        priority:
          type: string
        count:
          type: integer
      required:
        - priority
        - count

    TrendPoint:
      type: object
      properties:
        date:
          type: string
        created:
          type: integer
        resolved:
          type: integer
      required:
        - date
        - created
        - resolved