# lib/api-zod/src/generated/types/issue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for ML Issues Dashboard
 * OpenAPI spec version: 0.1.0
 */

export interface Issue {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  priority: string;
  status: string;
  /** @nullable */
  responsible?: string | null;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
}