# lib/api-zod/src/generated/types/issueStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for ML Issues Dashboard
 * OpenAPI spec version: 0.1.0
 */

export interface IssueStats {
  total: number;
  open: number;
  inProgress: number;
  resolved: number;
  critical: number;
}