# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for ML Issues Dashboard
 * OpenAPI spec version: 0.1.0
 */

export interface TrendPoint {
  date: string;
  created: number;
  resolved: number;
}