# lib/api-zod/src/generated/types/updateIssueInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for ML Issues Dashboard
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateIssueInput {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  responsible?: string | null;
  /** @nullable */
  dueDate?: string | null;
}